pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "marketlab"
version = "0.2.0"
description = "Reproducible market research toolkit for weekly modeling, walk-forward evaluation, and baseline-plus-ML experiments."
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
keywords = ["market-research", "backtesting", "walk-forward", "machine-learning", "portfolio"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Financial and Insurance Industry",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Topic :: Office/Business :: Financial :: Investment",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "anthropic>=0.49",
  "matplotlib>=3.8",
  "openai>=1.0",
  "pandas>=2.2",
  "PyYAML>=6.0",
  "yfinance>=0.2",
  "scikit-learn>=1.4",
  "scipy>=1.11",
]

[project.optional-dependencies]
dev = [
  "build>=1.2",
  "mcp>=1.27,<2",
  "mkdocs>=1.6",
  "pytest>=8.0",
  "ruff>=0.11",
  "tox>=4.0",
]
mcp = [
  "mcp>=1.27,<2",
]

[project.urls]
Homepage = "https://github.com/ricardogr07/market-lab"
Repository = "https://github.com/ricardogr07/market-lab"
Issues = "https://github.com/ricardogr07/market-lab/issues"

[dependency-groups]
dev = [
  "build>=1.2",
  "mcp>=1.27,<2",
  "mkdocs>=1.6",
  "pytest>=8.0",
  "ruff>=0.11",
  "tox>=4.0",
]

[project.scripts]
marketlab = "marketlab.cli:main"
marketlab-mcp = "marketlab.mcp.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.package-data]
"marketlab.resources.config_templates" = ["*.yaml"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
default-groups = ["dev"]

[tool.ruff]
target-version = "py312"
src = ["src", "tests", "scripts"]
extend-exclude = ["artifacts", "build", "dist", "site", ".pytest_tmp*", "tmpk_x17fq8"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--basetemp=.pytest_tmp -p no:cacheprovider"
markers = [
  "real_data: opt-in tests that download real market data.",
  "network: tests that depend on external network access.",
]
